Module snowbridge_system::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

create_agent

Sends a command to the Gateway contract to instantiate a new agent contract representing origin.

create_channel

Sends a message to the Gateway contract to create a new channel representing origin

force_transfer_native_from_agent

Sends a message to the Gateway contract to transfer ether from an agent to recipient.

force_update_channel

Sends a message to the Gateway contract to update an arbitrary channel

set_operating_mode

Sends a message to the Gateway contract to change its operating mode

set_pricing_parameters

Set pricing parameters on both sides of the bridge

set_token_transfer_fees

Sends a message to the Gateway contract to update fee related parameters for token transfers.

transfer_native_from_agent

Sends a message to the Gateway contract to transfer ether from an agent to recipient.

update_channel

Sends a message to the Gateway contract to update a channel configuration

upgrade

Sends command to the Gateway contract to upgrade itself with a new implementation contract