Trait snowbridge_system::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type OutboundQueue: SendMessage<Balance = BalanceOf<Self>>;
    type SiblingOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = MultiLocation>;
    type AgentIdOf: ConvertLocation<AgentId>;
    type Token: Mutate<Self::AccountId>;
    type TreasuryAccount: Get<Self::AccountId>;
    type DefaultPricingParameters: Get<PricingParametersOf<Self>>;
    type InboundDeliveryCost: Get<BalanceOf<Self>>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type OutboundQueue: SendMessage<Balance = BalanceOf<Self>>

Send messages to Ethereum

type SiblingOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = MultiLocation>

Origin check for XCM locations that can create agents

type AgentIdOf: ConvertLocation<AgentId>

Converts MultiLocation to AgentId

type Token: Mutate<Self::AccountId>

Token reserved for control operations

type TreasuryAccount: Get<Self::AccountId>

TreasuryAccount to collect fees

type DefaultPricingParameters: Get<PricingParametersOf<Self>>

Number of decimal places of local currency

type InboundDeliveryCost: Get<BalanceOf<Self>>

Cost of delivering a message from Ethereum

type WeightInfo: WeightInfo